refactor: restructure into src/core/ + src/adapters/rest/ (core/adapter seam)
Purely mechanical moves (git mv) with import rewrites — no behavior change:
- src/extractors/{snapshot,node-walker,built-in,types,finalize,index} -> src/core/
- src/transformers/* -> src/core/transformers/
- src/extractors/rest-*.ts -> src/adapters/rest/ (rest- prefix dropped)
- src/extractors/design-extractor.ts -> src/adapters/rest/design-extractor.ts

The core barrel (src/core/index.ts) is now the single pure entry Phase 2's
esbuild purity probe will target: it drops the adapter-owned
simplifyRawFigmaObject re-export and adds finalizeDesign so the whole core
is reachable from it. The Figma-free reachability gate reroots there.
Public library surface (src/index.ts) exports are unchanged.
