+++ src/core/finalize.ts
