+++ src/core/transformers/component.ts
